Refactoring to support longs in y19d9.

+++ src/2019/buri/aoc/y19/d02/Day02Test.java
@@ -13,13 +13,13 @@ public class Day02Test {
 
 	@Test
 	public void testGetInput() {
-		List<Integer> input = Day02.getInput(0);
+		List<Long> input = Day02.getInput(0);
 		assertEquals(157, input.size());
 	}
 
 	@Test
 	public void testPart1Examples() {
-		assertEquals(3500, Day02.getPart1Result(Day02.getInput(1), true));
+		assertEquals(3500L, Day02.getPart1Result(Day02.getInput(1), true));
 	}
 
 	/**
@@ -27,9 +27,9 @@ public class Day02Test {
 	 */
 	@Test
 	public void testPart1Puzzle() {
-		int result = Day02.getPart1Result(Day02.getInput(0), false);
+		long result = Day02.getPart1Result(Day02.getInput(0), false);
 		System.out.println("Day 2 Part 1\n\t" + result);
-		assertEquals(4930687, result);
+		assertEquals(4930687L, result);
 	}
 
 	/**
@@ -37,7 +37,7 @@ public class Day02Test {
 	 */
 	@Test
 	public void testPart2Puzzle() {
-		int result = Day02.getPart2Result(Day02.getInput(0));
+		long result = Day02.getPart2Result(Day02.getInput(0));
 		System.out.println("Day 2 Part 2\n\t" + result);
 		assertEquals(5335, result);
 	}
