+++ src/2019/buri/aoc/y19/d05/Day05Test.java
@@ -15,7 +15,7 @@ public class Day05Test {
 
 	@Test
 	public void testGetInput() {
-		List<Integer> input = Day05.getInput(0);
+		List<Long> input = Day05.getInput(0);
 		assertEquals(678, input.size());
 	}
 
@@ -24,14 +24,14 @@ public class Day05Test {
 	 */
 	@Test
 	public void testPart1Puzzle() {
-		int result = Day05.getResult(Part.ONE, Day05.getInput(0));
+		long result = Day05.getResult(Part.ONE, Day05.getInput(0));
 		System.out.println("Day 5 Part 1\n\t" + result);
-		assertEquals(15259545, result);
+		assertEquals(15259545L, result);
 	}
 
 	@Test
 	public void testPart2Examples() {
-		assertEquals(999, Day05.getResult(Part.ONE, Day05.getInput(1)));
+		assertEquals(999L, Day05.getResult(Part.ONE, Day05.getInput(1)));
 	}
 	
 	/**
@@ -39,8 +39,8 @@ public class Day05Test {
 	 */
 	@Test
 	public void testPart2Puzzle() {
-		int result = Day05.getResult(Part.TWO, Day05.getInput(0));
+		long result = Day05.getResult(Part.TWO, Day05.getInput(0));
 		System.out.println("Day 5 Part 2\n\t" + result);
-		assertEquals(7616021, result);
+		assertEquals(7616021L, result);
 	}
 }
